[FEATURE] user and group idstart integer
for a definded rage of integer for user id and group id i added the feature, that at the time you provision the ipa server you can define the range.
Somethimes this feature is very helpfull, because if you do not setup the idstart integer, the ipa-server-install script make a random choice, and other static dependecies will fail.

This parameter:

idstart

is used by calling the class "ipa".

Refs: #7
Status: ready