Rework admin user kinit handling
For now we require to know admin user id and we can know this id by
specifying idstart for ipa-server-install. Admin is always first
user to be created so it'll always get idstart uid and gid. The
idstart number is generated same way as it is by installer - it is
(2^31/200000)+10000
Also we generate keytab for user and kinit directly to default admin's
krb5cc cache which now is better implemented by not using tmp files in
Fedora20+ and RHEL7+.